// backend/src/films/dto/films.dto.ts
export class FilmDto {
  id: string;
  rating: number;
  director: string;
  tags: string[];
  title: string;
  about: string;
  description: string;
  image: string;
  cover: string;
}

export class ScheduleItemDto {
  id: string;
  daytime: string;
  hall: number;
  rows: number;
  seats: number;
  price: number;
  taken: string[];
}

export class FilmsListDto {
  total: number;
  items: FilmDto[];
}

export class ScheduleListDto {
  total: number;
  items: ScheduleItemDto[];
}